switched to float32

+++ sandbox/rocky/tf/algos/maml_il.py
@@ -56,20 +56,20 @@ class MAMLIL(BatchMAMLPolopt):
             ))
             adv_vars.append(tensor_utils.new_tensor(
                     'advantage' + stepnum + '_' + str(i),
-                    ndim=1, dtype=tf.float64,
+                    ndim=1, dtype=tf.float32,
                 ))
             if self.metalearn_baseline:
                 rewards_vars.append(tensor_utils.new_tensor(
                     'rewards' + stepnum + '_' + str(i),
-                    ndim=1, dtype=tf.float64,
+                    ndim=1, dtype=tf.float32,
                 ))
                 returns_vars.append(tensor_utils.new_tensor(
                     'returns' + stepnum + '_' + str(i),
-                    ndim=1, dtype=tf.float64,
+                    ndim=1, dtype=tf.float32,
                 ))
                 # path_lengths_vars.append(tensor_utils.new_tensor(
                 #     'path_lengths' + stepnum + '_' + str(i),
-                #     ndim=1, dtype=tf.float64,
+                #     ndim=1, dtype=tf.float32,
                 # ))
             expert_action_vars.append(self.env.action_space.new_tensor_variable(
                 name='expert_actions' + stepnum + '_' + str(i),
@@ -91,7 +91,7 @@ class MAMLIL(BatchMAMLPolopt):
         old_dist_info_vars, old_dist_info_vars_list = [], []
         for i in range(self.meta_batch_size):
             old_dist_info_vars.append({
-                k: tf.placeholder(tf.float64, shape=[None] + list(shape), name='old_%s_%s' % (i, k))
+                k: tf.placeholder(tf.float32, shape=[None] + list(shape), name='old_%s_%s' % (i, k))
                 for k, shape in dist.dist_info_specs
                 })
             old_dist_info_vars_list += [old_dist_info_vars[i][k] for k in dist.dist_info_keys]
@@ -99,7 +99,7 @@ class MAMLIL(BatchMAMLPolopt):
         theta0_dist_info_vars, theta0_dist_info_vars_list = [], []
         for i in range(self.meta_batch_size):
             theta0_dist_info_vars.append({
-                k: tf.placeholder(tf.float64, shape=[None] + list(shape), name='theta0_%s_%s' % (i, k))
+                k: tf.placeholder(tf.float32, shape=[None] + list(shape), name='theta0_%s_%s' % (i, k))
                 for k, shape in dist.dist_info_specs
                 })
             theta0_dist_info_vars_list += [theta0_dist_info_vars[i][k] for k in dist.dist_info_keys]
@@ -107,7 +107,7 @@ class MAMLIL(BatchMAMLPolopt):
         theta_l_dist_info_vars, theta_l_dist_info_vars_list = [], []  #theta_l is the current beta step's pre-inner grad update params
         for i in range(self.meta_batch_size):
             theta_l_dist_info_vars.append({
-                k: tf.placeholder(tf.float64, shape=[None] + list(shape), name='theta_l_%s_%s' % (i, k))
+                k: tf.placeholder(tf.float32, shape=[None] + list(shape), name='theta_l_%s_%s' % (i, k))
                 for k, shape in dist.dist_info_specs
                 })
             theta_l_dist_info_vars_list += [theta_l_dist_info_vars[i][k] for k in dist.dist_info_keys]
@@ -143,7 +143,7 @@ class MAMLIL(BatchMAMLPolopt):
 
             al_const = tf.constant(np.arange(self.max_path_length).reshape(-1, 1) / 100.0)
             al = tf.tile(al_const, tf.stack([tf.cast(tf.shape(obs_vars[0])[0] / self.max_path_length, tf.int32), 1]))
-            al = tf.cast(al, dtype=tf.float64)
+            al = tf.cast(al, dtype=tf.float32)
 
             for i in range(self.meta_batch_size):  # for training task T_i
                 adv = adv_vars[i]
@@ -242,7 +242,7 @@ class MAMLIL(BatchMAMLPolopt):
 
         outer_surr_obj = tf.reduce_mean(tf.stack(outer_surr_objs, 0))  # mean over all the different tasks
         input_vars_list += obs_vars + action_vars + expert_action_vars + old_dist_info_vars_list  # +adv_vars # TODO: kill action_vars from this list, and if we're not doing kl, kill old_dist_info_vars_list too
-        mean_kl = tf.cast(tf.reduce_mean(tf.concat(kls, 0)),tf.float64)
+        mean_kl = tf.cast(tf.reduce_mean(tf.concat(kls, 0)),tf.float32)
         #
         # # CORRECTION TERM ATTEMPT 2
         # term1_list = []
